nodetask: consolidate podCycle back into replace-pod (its only user)
podCycle was extracted to share the StatefulSet-fetch / owned-pod / delete
helpers between replace-pod and restart-pod. With RestartPod removed, replace-pod
is the sole user, so the shared abstraction is no longer justified
(CLAUDE.md: no premature helpers). Dissolve it:

- fetchStatefulSet/ownedPods/deletePod become methods on replacePodExecution
  (which now holds cfg ExecutionConfig directly, no embedded podCycle);
  guardSelectorAndReplicas/ownedByStatefulSet are plain unexported funcs.
- Delete pod_cycle.go / pod_cycle_test.go; the still-relevant unit tests move
  into replace_pod_test.go and the leftover restart-named fixtures are renamed
  to replace-pod equivalents.
- Drop podReady (dead — only restart-pod used it).

Pure internal refactor: replace-pod's revision-gated, readiness-blind behavior
is byte-for-byte preserved; no API/CRD change.
